books_project/requirements.txt
Django==5.0.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
duplicity==0.8.21
email-validator==1.3.1
exceptiongroup==1.1.2
Flask==2.2.5
Flask-AppBuilder==4.1.4
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
hdfs==2.7.0
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
idna==3.3
importlib-metadata==4.6.4
importlib-resources==6.0.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
json-merge-patch==0.2
jsonpath-ng==1.5.3
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==23.5.0
kubernetes==26.1.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
looker-sdk==23.12.0
louis==3.20.0
lxml==4.9.3
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
multidict==6.0.4
mypy-boto3-appflow==1.28.0
mypy-boto3-rds==1.28.11
mypy-boto3-redshift-data==1.28.0
netifaces==0.11.0
numpy==1.24.4
oauthlib==3.2.0
objsize==0.6.1
olefile==0.46
ordered-set==4.1.0
orjson==3.9.2
packaging==23.1
pandas==2.0.3
pandas-gbq==0.19.2
paramiko==2.9.3
pathspec==0.9.0
pendulum==2.1.2
pexpect==4.8.0
pg8000==1.29.2
Pillow==9.0.1
platformdirs==3.1.0
pluggy==1.2.0
ply==3.11
prison==0.2.1
proto-plus==1.22.3
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.2
psycopg2-binary==2.9.6
ptyprocess==0.7.0
py==1.10.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydantic==1.10.12
pydata-google-auth==1.8.1
pydot==1.4.2
Pygments==2.15.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.4.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyotp==2.9.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==7.2.2
python-amazon-sp-api==0.18.0
python-apt==2.4.0+ubuntu1
python-daemon==3.0.1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-dotenv==0.19.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2022.7.1
pytzdata==2020.1
pyxdg==0.27
PyYAML==5.4.1
redshift-connector==2.0.913
referencing==0.30.0
regex==2023.6.3
reportlab==3.6.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.1
scramp==1.4.4
SecretStorage==3.3.1
setproctitle==1.3.2
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlalchemy-bigquery==1.7.0
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
ssh-import-id==5.11
systemd-python==234
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
texttable==1.6.4
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
uc-micro-py==1.0.2
ufw==0.36.1
unattended-upgrades==0.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
watchtower==2.0.1
websocket-client==1.2.3
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
xdg==5
xkit==0.0.0
yarl==1.9.2
zipp==1.0.0
zstandard==0.21.0
